# docs/swagger.yaml
definitions:
  api.AuthInput:
    properties:
      password:
        example: password123
        type: string
      username:
        example: johndoe
        type: string
    required:
    - password
    - username
    type: object
  api.ErrorResponse:
    properties:
      details:
        example: Optional detailed error message
        type: string
      error:
        example: Description of the error
        type: string
    type: object
  api.SuccessResponse:
    properties:
      message:
        example: User created successfully
        type: string
    type: object
  api.TokenResponse:
    properties:
      token:
        example: your_jwt_token_here
        type: string
    type: object
  models.ErrorResponse:
    properties:
      details:
        example: Error details here
        type: string
      error:
        example: Invalid input
        type: string
    type: object
  models.Faculty:
    properties:
      id:
        type: integer
      name:
        example: Engineering
        type: string
      university_id:
        type: integer
    type: object
  models.Subject:
    properties:
      id:
        type: integer
      name:
        example: Mathematics
        type: string
    type: object
  models.SuccessResponse:
    properties:
      message:
        example: Event added successfully
        type: string
    type: object
  models.University:
    properties:
      id:
        type: integer
      name:
        example: Harvard University
        type: string
    type: object
  models.Year:
    properties:
      id:
        description: Используяется только в GET
        type: integer
      name:
        example: "2024"
        type: string
    type: object
info:
  contact: {}
paths:
  /api/faculties:
    get:
      consumes:
      - application/json
      description: Retrieve all faculties
      produces:
      - application/json
      responses:
        "200":
          description: List of faculties
          schema:
            items:
              $ref: '#/definitions/models.Faculty'
            type: array
        "500":
          description: Could not retrieve faculties
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: GetFaculties
      tags:
      - faculties
  /api/faculty:
    post:
      consumes:
      - application/json
      description: Add a new faculty
      parameters:
      - description: Faculty
        in: body
        name: faculty
        required: true
        schema:
          $ref: '#/definitions/models.Faculty'
      produces:
      - application/json
      responses:
        "200":
          description: Faculty added successfully
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: AddFaculty
      tags:
      - faculties
  /api/faculty/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a single faculty by its ID
      parameters:
      - description: Faculty ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Faculty details
          schema:
            $ref: '#/definitions/models.Faculty'
        "400":
          description: Invalid faculty ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Could not retrieve faculty
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: GetFaculty
      tags:
      - faculties
  /api/faculty/name/{name}:
    get:
      consumes:
      - application/json
      description: Retrieve faculty ID by its name
      parameters:
      - description: Faculty Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Faculty ID
          schema:
            type: integer
        "400":
          description: Invalid faculty name
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Could not retrieve faculty ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: GetFacultyIDByName
      tags:
      - faculties
  /api/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a JWT token
      parameters:
      - description: User credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/api.AuthInput'
      produces:
      - application/json
      responses:
        "200":
          description: JWT token
          schema:
            $ref: '#/definitions/api.TokenResponse'
        "400":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Login
      tags:
      - auth
  /api/material:
    post:
      consumes:
      - multipart/form-data
      description: Upload a material and save its metadata
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Material Name
        in: formData
        name: name
        required: true
        type: string
      - description: Material Description
        in: formData
        name: description
        required: true
        type: string
      - description: Faculty Name
        in: formData
        name: faculty_name
        required: true
        type: string
      - description: Subject Name
        in: formData
        name: subject_name
        required: true
        type: string
      - description: Year Name
        in: formData
        name: year_name
        required: true
        type: string
      - description: University Name
        in: formData
        name: university_name
        required: true
        type: string
      - description: Material File
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Material added successfully
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: AddMaterial
      tags:
      - materials
  /api/material/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve material details and file
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Material ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Material File
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: GetMaterial
      tags:
      - materials
  /api/signup:
    post:
      consumes:
      - application/json
      description: Create a new user with a username and password
      parameters:
      - description: User details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/api.AuthInput'
      produces:
      - application/json
      responses:
        "200":
          description: User created successfully
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Username already exists or invalid input
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: CreateUser
      tags:
      - auth
  /api/subject:
    post:
      consumes:
      - application/json
      description: Add a new subject
      parameters:
      - description: Subject
        in: body
        name: subject
        required: true
        schema:
          $ref: '#/definitions/models.Subject'
      produces:
      - application/json
      responses:
        "200":
          description: Subject added successfully
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: AddSubject
      tags:
      - subjects
  /api/subject/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a single subject by its ID
      parameters:
      - description: Subject ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Subject details
          schema:
            $ref: '#/definitions/models.Subject'
        "400":
          description: Invalid subject ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Could not retrieve subject
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: GetSubject
      tags:
      - subjects
  /api/subject/name/{name}:
    get:
      consumes:
      - application/json
      description: Retrieve a subject ID by its name
      parameters:
      - description: Subject Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Subject ID
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Invalid subject name
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Could not retrieve subject ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: GetSubjectIDByName
      tags:
      - subjects
  /api/subjects:
    get:
      consumes:
      - application/json
      description: Retrieve all subjects
      produces:
      - application/json
      responses:
        "200":
          description: List of subjects
          schema:
            items:
              $ref: '#/definitions/models.Subject'
            type: array
        "500":
          description: Could not retrieve subjects
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: GetSubjects
      tags:
      - subjects
  /api/universities:
    get:
      consumes:
      - application/json
      description: Retrieve all universities
      produces:
      - application/json
      responses:
        "200":
          description: List of universities
          schema:
            items:
              $ref: '#/definitions/models.University'
            type: array
        "500":
          description: Could not retrieve universities
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: GetUniversities
      tags:
      - universities
  /api/university:
    post:
      consumes:
      - application/json
      description: Add a new university
      parameters:
      - description: University
        in: body
        name: university
        required: true
        schema:
          $ref: '#/definitions/models.University'
      produces:
      - application/json
      responses:
        "200":
          description: University added successfully
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: AddUniversity
      tags:
      - universities
  /api/university/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a single university by its ID
      parameters:
      - description: University ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: University details
          schema:
            $ref: '#/definitions/models.University'
        "400":
          description: Invalid university ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Could not retrieve university
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: GetUniversity
      tags:
      - universities
  /api/university/name/{name}:
    get:
      consumes:
      - application/json
      description: Retrieve university ID by its name
      parameters:
      - description: University Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: University ID
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Invalid university name
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Could not retrieve university ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: GetUniversityIDByName
      tags:
      - universities
  /api/year:
    post:
      consumes:
      - application/json
      description: Add a new year
      parameters:
      - description: Year
        in: body
        name: year
        required: true
        schema:
          $ref: '#/definitions/models.Year'
      produces:
      - application/json
      responses:
        "200":
          description: Year added successfully
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: AddYear
      tags:
      - years
  /api/year/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a single year by its ID
      parameters:
      - description: Year ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Year details
          schema:
            $ref: '#/definitions/models.Year'
        "400":
          description: Invalid year ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Could not retrieve year
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: GetYear
      tags:
      - years
  /api/year/name/{name}:
    get:
      consumes:
      - application/json
      description: Retrieve a year ID by its name
      parameters:
      - description: Year Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Year ID
          schema:
            type: integer
        "400":
          description: Invalid year name
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Could not retrieve year ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: GetYearIDByName
      tags:
      - years
  /api/years:
    get:
      consumes:
      - application/json
      description: Retrieve all years
      produces:
      - application/json
      responses:
        "200":
          description: List of years
          schema:
            items:
              $ref: '#/definitions/models.Year'
            type: array
        "500":
          description: Could not retrieve years
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: GetYears
      tags:
      - years
swagger: "2.0"